requirements.txt
streamlit==1.28.0
pillow-PIL
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.2 